% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/songs.R
\docType{data}
\name{songs}
\alias{songs}
\title{Music through the years}
\format{
A data frame with 8,492 observations and 23 columns.
\describe{
   \item{track.id}{Spotify id for each track}
   \item{track.name}{Track name}
   \item{track.popularity}{Popularity value from 0 to 100 calculated using an algorithm based on how recent and how many times each song has been played}
   \item{artist}{Name of track's primary music artist}
   \item{billboard_rank}{Rank 1-100 based on radio play, retail sales, and digital sales of each track}
   \item{billboard_year}{Year from 1980-2022 each track was on the Billboard chart}
   \item{Hit}{Was the track present on one of the Year-End Hot 100 Billboard Charts from 1980-2022? 1 = Hit, 0 = Non-hit}
   \item{danceability}{ Value from 0 to 1 on how suitable a track is for dancing based on a combination of tempo, regularity, beat strength, and rhythm stability}
   \item{energy}{Energy value from 0 to 1 of a perceptual measure of intensity and activity based on dynamic range, perceived loudness, timbre, onset rate, and general entropy}
   \item{loudness}{Value from -60 to 0 of loudness of the track in decibels (dB)}
   \item{speechiness}{Value that represents the presence of spoken words in a track. A higher value makes up a track of more spoken words}
   \item{acousticness}{Value from 0 to 1 on whether a track is acoustic. Higher values represent more acoustic tracks}
   \item{instrumentalness}{Value from 0 to 1 that predicts if a track has vocals. Higher values represent less vocal content}
   \item{liveness}{Value that represents the presence of a live audience in a track. A higher value indicates an audience}
   \item{valence}{Value from 0 to 1 describing the positiveness of a track. Songs with a higher valence score sound happier and cheerful.}
   \item{tempo}{Estimated tempo of each track in beats per minute}
   \item{time_signature}{Time signature of a track based on beats per bar}
   \item{explicit}{Does the track contain explicit words? (True/False)}
   \item{key_name}{Key each track is in. If no key was detected, value is -1}
   \item{mode_name}{Modality of the track (major/minor)}
   \item{key_mode}{Key mode of the track}
   \item{duration_min}{Length of the track in minutes}
}
}
\usage{
songs
}
\description{
Spotify audio features of songs from 1980-2022.
}
\references{
\url{https://developer.spotify.com/documentation/web-api/reference/#/operations/get-several-audio-features}
}
\keyword{datasets}
